update to use real implementation and a mock implementation for dependencies

+++ src/main/scala/hellosubcut/Geocoder.scala
@@ -2,6 +2,6 @@ package hellosubcut
 
 trait Geocoder {
 
-  def getCoordinates(location: String): Option[(Float, Float)]
+  def getCoordinates(location: String): Option[(Double, Double)]
 
 }
